Created property-tree-utils file and moved some of the logic from property-tree into it. Manually tested - OK
